Busca de eventos conectados com db, mas com problema de não filtrar resultados encontrados. Ecra de salpico corrigido. Toast na tela de login.

+++ app/src/main/java/com/nearbysocialevents/nearby975/ActivityEvento.java
@@ -121,7 +121,7 @@ public class ActivityEvento extends Activity{
 
         txtNomeEvento.setText(event.nome);
         txtDescricaoEvento.setText(event.descricao);
-        txtPrecoEvento.setText("R$"+Float.toString(event.preco));
+        txtPrecoEvento.setText("R$"+Double.toString(event.preco));
         txtEnderecoEvento.setText(event.local);
         txtHoraEvento.setText(DateFormat.format("dd/MM/yyyy - hh:mm",event.data).toString());
